+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/field/index/fulltext/FulltextIndexModel.groovy
@@ -1,0 +1,22 @@
+package ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.field.index.fulltext
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 15.06.2015
+ */
+class FulltextIndexModel {
+
+    @FulltextIndex
+    String defaults
+
+    @FulltextIndex(useHashIndex = true)
+    String hash
+
+    @FulltextIndex(name = "all_options",
+            indexRadix = false,
+            ignoreChars = "'",
+            separatorChars = "!?",
+            minWordLength = 5,
+            stopWords = ["of", "the"])
+    String options
+}