+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/field/index/fulltext/FulltextIndexRemoteTest.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.field.index.fulltext
+
+import ru.vyarus.guice.persist.orient.util.remoteext.UseRemote
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 15.06.2015
+ */
+@UseRemote
+class FulltextIndexRemoteTest extends FulltextIndexTest {
+}