+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/field/index/lucene/LuceneIndexRemoteTest.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.guice.persist.orient.db.scheme.initializer.ext.field.index.lucene
+
+import ru.vyarus.guice.persist.orient.util.remoteext.UseRemote
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 16.06.2015
+ */
+@UseRemote
+class LuceneIndexRemoteTest extends LuceneIndexTest {
+}